Created a new class CommandGenerator in command_generator.py and moved generate_* functions from main to new class
