+++ src/main.py
@@ -8,6 +8,8 @@ import serializer
 import deserializer
 import utils
 
+import command_generator
+
 # Serialize the contents of base.json into a dictionary
 class Main:
     """
@@ -21,6 +23,14 @@ class Main:
         self.start_commands = serialize.get_start_commands()
         self.end_commands = serialize.get_end_commands()
         self.config = serialize.get_config()
+
+        self.command_generator = command_generator.CommandGenerator()
+        # Create instances of the generate_* methods
+        self.generate_first_name_command = command_generator.generate_first_name_command()
+        self.generate_last_name_command = command_generator.generate_last_name_command()
+        self.generate_email_command = command_generator.generate_email_command()
+        self.generate_save_and_add_command = command_generator.generate_save_and_add_command()
+        self.generate_pause_command = command_generator.generate_pause_command()
 
         self.base['id'] = utils.generate_uuid()
         self.base['name'] = self.config['event_name']
@@ -40,90 +50,6 @@ class Main:
         Sets test commands to start commands in the base object
         """
         self.base['tests'][0]['commands'] = self.start_commands
-
-    def generate_first_name_command(self, first_name):
-        """
-        Generates type first name command
-        """
-        command = copy.deepcopy(self.command_template)
-        command['id'] = utils.generate_uuid()
-        command['command'] = 'type'
-        command['target'] = 'id=s1vcua'
-        command['targets'] = [
-            ["id=s1vcua", "id"],
-            ["name=first_name", "name"],
-            ["css=#s1vcua", "css:finder"],
-            ["xpath=//input[@id='s1vcua']", "xpath:attributes"],
-            ["xpath=//div[@id='overlay-container']/div/div/div[2]/form/div/div/div/div/div/input", "xpath:idRelative"],
-            ["xpath=//form/div/div/div/div/div/input", "xpath:position"]
-        ]
-        command['value'] = first_name
-        return command
-
-    def generate_last_name_command(self, last_name):
-        """
-        Generates type last name command
-        """
-        command = copy.deepcopy(self.command_template)
-        command['id'] = utils.generate_uuid()
-        command['command'] = 'type'
-        command['target'] = 'id=s1yx3r'
-        command['targets'] = [
-            ["id=s1yx3r", "id"],
-            ["name=last_name", "name"],
-            ["css=#s1yx3r", "css:finder"],
-            ["xpath=//input[@id='s1yx3r']", "xpath:attributes"],
-            ["xpath=//div[@id='overlay-container']/div/div/div[2]/form/div/div[2]/div/div/div/input", "xpath:idRelative"],
-            ["xpath=//div[2]/div/div/div/input", "xpath:position"]
-        ]
-        command['value'] = last_name
-        return command
-
-    def generate_email_command(self, email):
-        """
-        Generates type email command
-        """
-        command = copy.deepcopy(self.command_template)
-        command['id'] = utils.generate_uuid()
-        command['command'] = 'type'
-        command['target'] = 'id=s2v3vd'
-        command['targets'] = [
-            ["id=s2v3vd", "id"],
-            ["name=email", "name"],
-            ["css=#s2v3vd", "css:finder"],
-            ["xpath=//input[@id='s2v3vd']", "xpath:attributes"],
-            ["xpath=//div[@id='overlay-container']/div/div/div[2]/form/div/div[3]/div/div/div/input", "xpath:idRelative"],
-            ["xpath=//div[3]/div/div/div/input", "xpath:position"]
-        ]
-        command['value'] = email
-        return command
-
-    def generate_save_and_add_command(self):
-        """
-        Generates click command for Save and Add More
-        """
-        command = copy.deepcopy(self.command_template)
-        command['id'] = utils.generate_uuid()
-        command['command'] = 'click'
-        command['target'] = 'css=.Button-Button__outlined_bddio'
-        command['targets'] = [
-            ["css=.Button-Button__outlined_bddio", "css:finder"],
-            ["xpath=(//button[@type='button'])[5]", "xpath:attributes"],
-            ["xpath=//div[@id='overlay-container']/div/div/div[2]/form/div/div[6]/div/div[2]/div/button", "xpath:idRelative"],
-            ["xpath=//div[2]/div/button", "xpath:position"],
-            ["xpath=//button[contains(.,'Save and add more')]", "xpath:innerText"]
-        ]
-        return command
-    
-    def generate_pause_command(self, target=4000):
-        """
-        Generates command pausing the execution for the given number of seconds
-        """
-        command = copy.deepcopy(self.command_template)
-        command['id'] = utils.generate_uuid()
-        command['command'] = 'pause'
-        command['target'] = str(target)
-        return command
 
     def add_attendees_commands(self):
         """
